15주차(REAT_API,JS)/221018/ws_82/requirements.txt
asgiref==3.3.4
Django==3.2
django-seed==0.2.2
djangorestframework==3.12.4
Faker==8.1.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
asgiref==3.3.4
Django==3.2
django-extensions==3.2.1
django-seed==0.2.2
djangorestframework==3.12.4
Faker==8.1.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
